Move graphql-testx tests into tests folder
